AbsenceApplicationとWorkHistoryManagement周りの改修。主に、コントローラークラスのロジックの一部ををサービスクラスのメソッドに移したり、コードを整理したりした。

+++ src/main/java/com/example/webapp/service/impl/AbsenceApplicationServiceImpl.java
@@ -1,16 +1,19 @@
 package com.example.webapp.service.impl;
 
+import java.time.LocalDate;
 import java.util.List;
 
-import org.springframework.security.core.Authentication;
-import org.springframework.security.core.authority.SimpleGrantedAuthority;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.stereotype.Service;
 
 import com.example.webapp.entity.AbsenceApplication;
 import com.example.webapp.entity.AbsenceReason;
-import com.example.webapp.entity.Role;
+import com.example.webapp.exception.DuplicateApplicationException;
+import com.example.webapp.exception.InvalidAccessException;
+import com.example.webapp.exception.InvalidEditException;
+import com.example.webapp.form.AbsenceApplicationForm;
+import com.example.webapp.helper.AbsenceApplicationHelper;
 import com.example.webapp.repository.AbsenceApplicationMapper;
-import com.example.webapp.repository.EmployeesManagementMapper;
 import com.example.webapp.service.AbsenceApplicationService;
 
 import lombok.RequiredArgsConstructor;
@@ -20,7 +23,6 @@ import lombok.RequiredArgsConstructor;
 public class AbsenceApplicationServiceImpl implements AbsenceApplicationService {
 
 	private final AbsenceApplicationMapper absenceApplicationMapper;
-	private final EmployeesManagementMapper employeeManagementMapper;
 
 //	@Override
 //	public List<AbsenceApplication> selectAllApplicationsAfterToday() {
@@ -41,13 +43,18 @@ public class AbsenceApplicationServiceImpl implements AbsenceApplicationService 
 //	}
 //	
 	@Override
+	public List<AbsenceApplication> get(String fromPage,Integer employeeId) throws InvalidAccessException {
+		if (fromPage.equals("admin")) {
+			return absenceApplicationMapper.selectAll();
+		}
+		if(fromPage.equals("user")) {
+			return absenceApplicationMapper.selectAllByEmployeeId(employeeId);
+		}
+		throw new InvalidAccessException("遷移元ページが不明です");
+	}
+	@Override
 	public List<AbsenceApplication> getTodayApplications(){
 		return absenceApplicationMapper.selectToday();
-	}
-	
-	@Override
-	public List<AbsenceApplication> getApplicationDetail(){
-		return 	absenceApplicationMapper.selectByApplicationId();
 	}
 
 	@Override
@@ -56,32 +63,27 @@ public class AbsenceApplicationServiceImpl implements AbsenceApplicationService 
 	}
 
 	@Override
-	public void insertApplication(AbsenceApplication application) {
+	public void submitApplication(AbsenceApplicationForm form) throws DuplicateApplicationException {
+		AbsenceApplication application = AbsenceApplicationHelper.convertAbsenceApplication(form);
+		try {
 		absenceApplicationMapper.insert(application);
+		}catch(DataIntegrityViolationException e) {
+			throw new DuplicateApplicationException("すでに申請済みです");
+		}
 	}
 
 	@Override
 	public void updateApprove(Integer shiftId, Boolean decision) {
-		try {
 			absenceApplicationMapper.update(shiftId, decision);
-		} catch (Exception e) {
-			
-		}
 	}
 
 	@Override
-	public void deleteApplication(Integer applicationId) {
+	public void deleteApplication(Integer applicationId) throws InvalidEditException {
+		AbsenceApplication application=absenceApplicationMapper.selectByApplicationId(applicationId);
+		Boolean isPast=application.getShiftSchedule().getDate().isBefore(LocalDate.now());
+		if(isPast) {
+			throw new InvalidEditException("過去の申請は削除できません");
+		}
 		absenceApplicationMapper.delete(applicationId);
-	}
-
-	@Override
-	public List<AbsenceApplication> get(Authentication auth) {
-		Boolean isAdmin=auth.getAuthorities().contains(new SimpleGrantedAuthority(Role.ADMIN.toString()));
-		if (isAdmin) {
-			return absenceApplicationMapper.selectAll();
-		} else {
-			Integer employeeId = Integer.parseInt(auth.getName());
-			return absenceApplicationMapper.selectAllByEmployeeId(employeeId);
-		}
 	}
 }
